refactor: Extract tokens from askinfo
